@@ -18,62 +18,63 @@
     "lint:code": "ts-standard",
     "commit": "lint-staged --allow-empty",
     "push": "run-s lint:es build",
-    "envsub": "envsub --diff ./projects/dapp/src/environments/environment.prod.ts"
+    "envsub": "envsub --env-file .env --all --diff ./projects/dapp/src/environments/environment.prod.ts",
+    "postinstall": "husky install"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~11.2.3",
-    "@angular/cdk": "^11.2.2",
-    "@angular/common": "~11.2.3",
-    "@angular/compiler": "~11.2.3",
-    "@angular/core": "~11.2.3",
-    "@angular/forms": "~11.2.3",
-    "@angular/material": "^11.2.2",
-    "@angular/platform-browser": "~11.2.3",
-    "@angular/platform-browser-dynamic": "~11.2.3",
-    "@angular/router": "~11.2.3",
-    "@ngneat/transloco": "^2.19.1",
+    "@angular/animations": "~11.2.4",
+    "@angular/cdk": "^11.2.3",
+    "@angular/common": "~11.2.4",
+    "@angular/compiler": "~11.2.4",
+    "@angular/core": "~11.2.4",
+    "@angular/forms": "~11.2.4",
+    "@angular/material": "^11.2.3",
+    "@angular/platform-browser": "~11.2.4",
+    "@angular/platform-browser-dynamic": "~11.2.4",
+    "@angular/router": "~11.2.4",
+    "@ngneat/transloco": "^2.20.1",
     "@ngxd/core": "^10.0.0",
     "@types/lodash": "^4.14.168",
-    "@waves.exchange/provider-web": "^0.2.5",
-    "@waves/signer": "0.1.2",
+    "@waves.exchange/provider-web": "^1.0.2",
+    "@waves/signer": "1.0.2",
     "envsub": "^4.0.7",
-    "eslint-config-prettier": "^7.2.0",
+    "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^3.3.1",
     "include-media": "^1.4.9",
-    "ngx-markdown": "^10.1.1",
-    "ngx-mask": "^11.0.1",
+    "ngx-markdown": "^11.1.1",
+    "ngx-mask": "^11.1.4",
     "ngx-progressbar": "^6.1.0",
     "npm-run-all": "^4.1.5",
-    "pm2": "^4.5.4",
+    "pm2": "^4.5.5",
     "prettier": "^2.2.1",
     "rxjs": "~6.6.6",
     "tslib": "^2.1.0",
     "tslint-etc": "^1.13.9",
-    "zone.js": "~0.10.2"
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1102.2",
-    "@angular-eslint/builder": "1.0.0",
-    "@angular-eslint/eslint-plugin": "1.0.0",
-    "@angular-eslint/eslint-plugin-template": "1.0.0",
-    "@angular-eslint/schematics": "1.0.0",
-    "@angular-eslint/template-parser": "1.0.0",
-    "@angular/cli": "~11.2.2",
-    "@angular/compiler-cli": "~11.2.3",
+    "@angular-devkit/build-angular": "~0.1102.3",
+    "@angular-eslint/builder": "1.2.0",
+    "@angular-eslint/eslint-plugin": "1.2.0",
+    "@angular-eslint/eslint-plugin-template": "1.2.0",
+    "@angular-eslint/schematics": "1.2.0",
+    "@angular-eslint/template-parser": "1.2.0",
+    "@angular/cli": "~11.2.3",
+    "@angular/compiler-cli": "~11.2.4",
     "@types/jasmine": "3.6.4",
     "@types/jasminewd2": "2.0.8",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "^4.14.1",
-    "@typescript-eslint/parser": "4.3.0",
-    "eslint": "^7.18.0",
+    "@types/node": "^14.14.31",
+    "@typescript-eslint/eslint-plugin": "^4.16.1",
+    "@typescript-eslint/parser": "4.16.1",
+    "eslint": "^7.21.0",
     "eslint-config-standard-with-typescript": "^20.0.0",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsdoc": "30.7.6",
+    "eslint-plugin-jsdoc": "32.2.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prefer-arrow": "1.2.2",
-    "eslint-plugin-promise": "^4.2.1",
-    "husky": "^4.3.8",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "eslint-plugin-promise": "^4.3.1",
+    "husky": "^5.1.3",
     "jasmine-core": "3.6.0",
     "jasmine-spec-reporter": "5.0.2",
     "karma": "5.2.3",
