@@ -1,24 +1,20 @@
-import { Inject, Injectable } from '@angular/core'
-import { Signer, IUserData } from '@waves/signer/'
-import Provider from '@waves.exchange/provider-web'
-import { API, AppApiInterface } from '@constants'
+import {Inject, Injectable, isDevMode} from '@angular/core'
+import {Signer, UserData} from '@waves/signer/'
+import {ProviderWeb} from '@waves.exchange/provider-web'
+import {API, AppApiInterface} from '@constants'
 import {
   SignerUser,
-  SignerInvokeArgs,
   TransactionRawState,
   TransactionState,
-  TransactionsSuccessResult
+  TransactionsSuccessResult, ISignerInvokeAnyData
 } from './signer.model'
-import { BehaviorSubject, from, Observable } from 'rxjs'
-import { publishReplay, refCount, tap, switchMap, retryWhen, delay, map, take } from 'rxjs/operators'
-import {
-  IInvoke, IInvokeWithType,
-  IMoney, TParamsToSign
-} from '@waves/signer/cjs/interface'
-import { HttpClient } from '@angular/common/http'
-import { translate } from '@ngneat/transloco'
-import { MatSnackBar } from '@angular/material/snack-bar'
-import { StorageService } from '@services/storage/storage.service'
+import {BehaviorSubject, from, Observable} from 'rxjs'
+import {publishReplay, refCount, tap, switchMap, retryWhen, delay, map, take} from 'rxjs/operators'
+import {HttpClient} from '@angular/common/http'
+import {translate} from '@ngneat/transloco'
+import {MatSnackBar} from '@angular/material/snack-bar'
+import {StorageService} from '@services/storage/storage.service'
+import {InvokeScriptCallArgument} from '@waves/ts-types/src/parts'
 
 @Injectable({
   providedIn: 'root'
@@ -26,7 +22,12 @@ import { StorageService } from '@services/storage/storage.service'
 export class SignerService {
   public readonly signer!: Signer
 
-  private readonly user$: BehaviorSubject<SignerUser> = new BehaviorSubject({ name: '', address: '', publicKey: '', balance: '' })
+  private readonly user$: BehaviorSubject<SignerUser> = new BehaviorSubject({
+    name: '',
+    address: '',
+    publicKey: '',
+    balance: ''
+  })
 
   public user: Observable<SignerUser> = this.user$.pipe(tap((data) => {
     this.storageService.userData = data
@@ -38,11 +39,16 @@ export class SignerService {
     private readonly snackBar: MatSnackBar,
     private storageService: StorageService
   ) {
+
     this.signer = new Signer({
       // Specify URL of the node on Testnet
       NODE_URL: api.nodes // eslint-disable-line
     })
-    this.signer.setProvider(new Provider(api.signer)).catch(() => {})
+    this.signer.setProvider(new ProviderWeb(api.signer)).catch((e) => {
+      if (isDevMode()) {
+        console.log('Error ProviderWeb:', e)
+      }
+    })
 
     this.user$.next(this.storageService.userData as SignerUser)
   }
@@ -50,9 +56,12 @@ export class SignerService {
   public login (): Observable<Observable<SignerUser>> {
     return from(
       this.signer.login()
-        .then((user: IUserData) => {
+        .then((user: UserData) => {
           this.signer.getBalance()
             .then((res) => {
+              if (isDevMode()) {
+                console.log('getBalance:', res)
+              }
               this.user$.next({
                 ...user,
                 balance: res[0].amount.toString(),
@@ -69,31 +78,50 @@ export class SignerService {
   }
 
   public logout (): Observable<void> {
-    this.user$.next({ name: '', address: '', publicKey: '', balance: '' })
+    this.user$.next({name: '', address: '', publicKey: '', balance: ''})
     return from(this.signer.logout())
   }
 
   public invokeProcess (
     contractAddress: string,
     command: string,
-    args: SignerInvokeArgs[],
-    payment: IMoney[] = []
+    args: Array<InvokeScriptCallArgument<string | number>>,
+    payment: {
+      assetId: string,
+      amount: number | string,
+    }[] = []
   ): Observable<TransactionsSuccessResult> {
+    if (isDevMode()) {
+      console.log('invokeProcess params contractAddress:', contractAddress)
+      console.log('invokeProcess params command:', command)
+      console.log('invokeProcess params args:', args)
+      console.log('invokeProcess params payment:', payment)
+    }
     return from(this.signer.invoke({
-      payment,
+      ...(payment && payment.length ? {
+        payment
+      } : {}),
       dApp: contractAddress,
       call: {
         function: command,
         args
-      }
-    } as IInvoke).sign()).pipe(
+      },
+      feeAssetId: null
+    }).sign()).pipe(
       take(1),
       tap(() => {
         this.snackBar.open(translate('messages.startTransaction'), translate('messages.ok'))
       }),
-      switchMap((tx: TParamsToSign<IInvokeWithType>) => from(this.signer.broadcast(tx))),
-      // @ts-expect-error: Data is a specific type on Signer library
-      switchMap((data) => this.status(data?.id)),
+      switchMap((tx) => from(this.signer.broadcast(tx).catch((e) => {
+        console.warn(e)
+        throw new Error('broadcast could not process the request:')
+      }))),
+      switchMap((data) => {
+        const d: ISignerInvokeAnyData = {
+          ...data
+        }
+        return this.status(d.id)
+      }),
       tap(() => {
         this.snackBar.open('Transaction is complete', translate('messages.ok'))
       })
@@ -106,13 +134,15 @@ export class SignerService {
       params: {
         id: tx
       },
-      headers: { accept: 'application/json; charset=utf-8' }
+      headers: {accept: 'application/json; charset=utf-8'}
     }).pipe(
       map((data: TransactionState[]) => {
         const confirmation = data.find((state: TransactionState) =>
           state.status === 'confirmed' && state.confirmations >= this.api.confirmations)
 
-        console.log('Confirmation', confirmation)
+        if(isDevMode()) {
+          console.log('Confirmation', confirmation)
+        }
         if (!confirmation) {
           throw new Error('wait')
         }
